// portal/src/types.ts
export type EyeConfig = {
  pupil_length: number;
  wiggle_amp: number;
  blink_interval: number;
  eye_red: number;
  eye_green: number;
  eye_blue: number;
  sound_threshold: number;
  sin_rarity: number;
  eyeroll_rarity: number;
  startled_rarity: number;
  suspicious_rarity: number;
  angry_rarity: number;
};